Move StandalonePersistRepository
