+++ shardingsphere-infra/shardingsphere-infra-common/src/main/resources/META-INF/services/org.apache.shardingsphere.infra.mode.impl.standalone.StandalonePersistRepository
